requirements.txt
aniso8601==9.0.1
apispec==6.3.0
attrs==23.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.3
flask-apispec==0.11.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
mistune==3.0.1
packaging==23.1
psycopg2==2.9.6
PyJWT==2.7.0
pyrsistent==0.19.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.9
typing_extensions==4.5.0
webargs==8.2.0
Werkzeug==2.2.3
